# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VektraHub SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile with role and business
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /me/onboard:
    post:
      operationId: onboardBusiness
      tags: [users]
      summary: Onboard a new business for the current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardInput"
      responses:
        "201":
          description: Business created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products for the current business
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        clerkId:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [client, admin]
        business:
          $ref: "#/components/schemas/Business"
      required:
        - clerkId
        - email
        - role
    UserProfileInput:
      type: object
      properties:
        name:
          type: string
    OnboardInput:
      type: object
      properties:
        businessName:
          type: string
        sector:
          type: string
          enum: [restaurante, kiosco, dietetica, ferreteria, ropa, distribuidora, otro]
        plan:
          type: string
          enum: [starter, pro]
      required:
        - businessName
        - sector
        - plan
    Business:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sector:
          type: string
        plan:
          type: string
          enum: [starter, pro]
        status:
          type: string
          enum: [active, suspended, trial]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - sector
        - plan
        - status
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        category:
          type: ["string", "null"]
        stock:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - price
        - active
        - createdAt
    ProductInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        stock:
          type: integer
        imageUrl:
          type: string
        active:
          type: boolean
      required:
        - name
        - price
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        stock:
          type: integer
        imageUrl:
          type: string
        active:
          type: boolean
...